==> notebooks/01_prepare_data_download.ipynb <==
## Download the data for all the tickers that are to be analyzed

import yfinance as yf
import pandas as pd
import sys
import os
from pathlib import Path

Get the list of tickers from universe.py

#Get list of tickers from src/universe.py

sys.path.append("../src")
from universe import get_tickers_list

tickers_list = get_tickers_list()

print(len(tickers_list))
print(tickers_list)

Downloading the raw data for eachticker and saving it under data/raw using the format ticker.csv

#make individual csv files of OHLCV data for all tickers

# Date range
start_date = "2022-12-01"
end_date = "2024-12-31"

failed_tickers = []
empty_tickers = []

#output directory path and create it if it doesn't exist
output_dir = Path("../data/raw")

#create dir if it doesn't exist
output_dir.mkdir(parents=True, exist_ok=True)


#download all the raw data and store it in the output location

for ticker in tickers_list:
    #chnaging naming convention for yahoo finance we need '-' in ticker name instead of '.' eg: BRK-B not BRK.B
    yf_ticker = ticker.replace('.', '-')
    try:
        #downlaod data ticker by ticker and store in an csv file and name the file ticker.csv
        df = yf.download(yf_ticker, start=start_date, end= end_date, auto_adjust= True)
        if not df.empty:
            df.to_csv(output_dir/f"{ticker}.csv")
            #print(f" Saved: {ticker}")
        
        else:
            #print(f" Empty data: {ticker}")
            empty_tickers.append(ticker)
    
    except Exception as e:
        ##print(f" Failed: {ticker} — {e}")
        failed_tickers.append(ticker)


# print(f"\nTotal tickers attempted: {len(tickers_list)}")
# print(f"Saved: {len(tickers_list) - len(failed_tickers) - len(empty_tickers)}")
# print(f"Empty: {len(empty_tickers)} → {empty_tickers}")
# print(f"Failed: {len(failed_tickers)} → {failed_tickers}")
