==> notebooks/02_clean_format_data.ipynb <==
## Saving only the necessarry data in new csv files 

import yfinance as yf
import pandas as pd
import sys
import os
from pathlib import Path

for each file in data/raw saving only necassry column and storing them in data/stock_data using the format ticker.csv

#input dir from where we will get out raw data
input_path = "../data/raw"

#output dir where we will store clean data
output_path = "../data/stock_data"


output_file = Path("../data/stock_data/AAPL.csv")


for file in os.listdir(input_path):
    try:
        file_path= Path(file)
        input_file = input_path / file_path
        output_file = output_path / file_path

        # Read and clean the CSV
        df = pd.read_csv(input_file, skiprows=2)
        df.columns = ["Date", "Close", "High", "Low", "Open", "Volume"]
        df = df[["Date", "Open", "High", "Low", "Close", "Volume"]]
        df["Date"] = pd.to_datetime(df["Date"])
        df = df.sort_values("Date").set_index("Date")

        # Save cleaned file
        df.to_csv(output_file)
        print(f"✅ Cleaned and saved: {output_file.name}")

    except Exception as e:
        print(f"⚠️ Skipped {file} due to error: {e}")




Checking if all the files were processed 

#path = "your_directory_path"
items = os.listdir(output_path)
print("Total:", len(items))
