==> notebooks/03_compute_momentum_features.ipynb <==
## Creating the momentum files

defineing the function that will calculate the momentum scores

import pandas as pd
from pathlib import Path
import os

#adding momentum scores columns 1month, 3 months, 6 months, 12 months
def add_momentum_columns(df):
    df["ret_1m"] = df["Close"]/df["Close"].shift(21)-1 #21 trading days in a month
    df["ret_3m"] = df["Close"]/df["Close"].shift(63)-1 #63 trading days in 3 months
    df["ret_6m"] = df["Close"]/df["Close"].shift(126)-1 #126 trading days in 6 months
    df["ret_12m"] = df["Close"]/df["Close"].shift(252)-1 #252 trading days in a year
    return df

For each file in data/stocks_data, calculate momentum scores and save the results (including initial data and momentum columns) to data/momentum_data.

#test for one file


input_path = Path("../data/stock_data")

output_path = Path("../data/momentum_data")

for file in input_path.iterdir():
    try:
        input_file = file
        output_file = output_path/file.name

        df = pd.read_csv(input_file, parse_dates=["Date"], index_col = "Date")
        df = df.sort_index() 
        df = add_momentum_columns(df)
        df = df[df.index>= "2024-01-01"]

        df.to_csv(output_file)
        print(f"Added momentum cols and saved: {output_file.name}")

    except Exception as e:
        print(f" skipped{file} due to error{e}")
