==> notebooks/06_metrics_calculations.ipynb <==
## Risk Metriccs Calculation

import pandas as pd
import numpy as np
import json

Path to data used for risk calculations

df = pd.read_csv("../data/backtest_results/monthly_performance.csv", parse_dates=["Date"])

df["return_pct"] = pd.to_numeric(df["return_pct"], errors="coerce")


**Cumulative Return**

highest total prfoit throughout the months

**Volitility**

standard deviation of returns%

**Annual Volitility**

annual volitility = monthly volitility * sqareroot(12)

**Sharpe Ratio**

Average monthly return per unit of risk (volatility)

**Drawdown**

current value compared to highest profit

**Calmar Ratio**

max returns/max loss = cummulative_max/|Max_drawdown|

#cummulatice return
cummulative_return=(df["portfolio_value"].iloc[-1]/df["portfolio_value"].iloc[0])-1

#monthly stats
mean_return= df["return_pct"].mean()
volatility_monthly= df["return_pct"].std()

#annualized metrics, standard formula monthly standard deviation times sqrt(12)
volatility_annual = volatility_monthly*np.sqrt(12)
sharpe_ratio = mean_return/volatility_monthly if volatility_monthly!=0 else np.nan

df["cum_max"] = df["portfolio_value"].cummax()
df["drawdown"] = df["portfolio_value"]/df["cum_max"]-1
max_drawdown = df["drawdown"].min()
calmar_ratio = cummulative_return/abs(max_drawdown) if max_drawdown != 0 else np.nan

Storing Risk metrics as a JSON file

risk_metrics = {
    "Cummulatice return(%)": round(cummulative_return*100,2),
    "Monthly Volatility (%)": round(volatility_monthly,2),
    "Annulized Volatility(%)": round(volatility_annual,2),
    "Sharpe ratio": round(sharpe_ratio,2),
    "Max Drawdown(%)": round(max_drawdown*100,2),
    "Calmar Ratio": round(calmar_ratio,2),
}


with open("../data/backtest_results/risk_metrics.json","w") as f:
    json.dump(risk_metrics,f,indent=4)


print("Risk metrics saved in backtest_results/risk_metrics.json")
